Update go-whosonfirst-export to 0.3.0
Updated go-whosonfirst-export to 0.3.0 so we're exporting our GeoJSON files
using the latest version. The update has changed the API slightly, so reworked
the setup in main.go to match. This builds, but the default ID provider can't
generate enough IDs for us, so syncing fails when the first request fails.